[ADD] Admin Controller & Course Controller

+++ src/main/java/com/dalc/one/controller/PlaceController.java
@@ -1,6 +1,7 @@
 package com.dalc.one.controller;
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -13,11 +14,14 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
+import com.dalc.one.ExceptionEnum;
 import com.dalc.one.domain.Place;
 import com.dalc.one.domain.User;
 import com.dalc.one.domain.UserLikePlace;
 import com.dalc.one.domain.UserSearchPlace;
+import com.dalc.one.jwt.JwtTokenProvider;
 import com.dalc.one.service.LehgoFacade;
 
 import lombok.RequiredArgsConstructor;
@@ -62,13 +66,32 @@ public class PlaceController{
 	/* 사용자 선호 여행지 리스트*/
 	@ResponseBody
 	@PostMapping("mylist")
-	public ResponseEntity<List<UserLikePlace>> getUserPlaceList(@Valid @RequestBody User user) throws Exception {
+	public ResponseEntity<List<UserLikePlace>> getUserPlaceList(HttpServletRequest request, @Valid @RequestBody User user) throws Exception {
+		String authorizationHeader = request.getHeader("authorization");
+		if (authorizationHeader == null) {
+			throw new ResponseStatusException
+				(ExceptionEnum.NOT_LOGIN.getStatus(), ExceptionEnum.NOT_LOGIN.getMessage());
+		}
+		else if (!JwtTokenProvider.getUserOf(authorizationHeader).getUsername().equals(user.getId())) {
+			throw new ResponseStatusException
+				(ExceptionEnum.NOT_MATCH.getStatus(), ExceptionEnum.NOT_LOGIN.getMessage());
+		}
 		return ResponseEntity.ok(lehgo.getUserPlaceList(user));
 	}
 	
 	@ResponseBody
 	@PostMapping("mylist/like")
-	public ResponseEntity<HttpStatus> addUserPlace(@Valid @RequestBody User user, @RequestParam("id") int id ) throws Exception {
+	public ResponseEntity<HttpStatus> addUserPlace(HttpServletRequest request, @Valid @RequestBody User user, @RequestParam("id") int id ) throws Exception {
+		String authorizationHeader = request.getHeader("authorization");
+		if (authorizationHeader == null) {
+			throw new ResponseStatusException
+				(ExceptionEnum.NOT_LOGIN.getStatus(), ExceptionEnum.NOT_LOGIN.getMessage());
+		}
+		else if (!JwtTokenProvider.getUserOf(authorizationHeader).getUsername().equals(user.getId())) {
+			throw new ResponseStatusException
+				(ExceptionEnum.NOT_MATCH.getStatus(), ExceptionEnum.NOT_LOGIN.getMessage());
+		}
+		
 		UserLikePlace result = lehgo.addUserplace(user, id);
 				
 		if (result != null) {
@@ -82,7 +105,17 @@ public class PlaceController{
 	
 	@ResponseBody
 	@PostMapping("mylist/delete")
-	public ResponseEntity<HttpStatus> deleteUserPlace(@Valid @RequestBody User user, @RequestParam("id") int id ) throws Exception {
+	public ResponseEntity<HttpStatus> deleteUserPlace(HttpServletRequest request, @Valid @RequestBody User user, @RequestParam("id") int id ) throws Exception {
+		String authorizationHeader = request.getHeader("authorization");
+		if (authorizationHeader == null) {
+			throw new ResponseStatusException
+				(ExceptionEnum.NOT_LOGIN.getStatus(), ExceptionEnum.NOT_LOGIN.getMessage());
+		}
+		else if (!JwtTokenProvider.getUserOf(authorizationHeader).getUsername().equals(user.getId())) {
+			throw new ResponseStatusException
+				(ExceptionEnum.NOT_MATCH.getStatus(), ExceptionEnum.NOT_LOGIN.getMessage());
+		}
+		
 		int result = lehgo.deleteUserPlace(user, id);
 
 		if(result > 0) {
